PBI-28: Implement check if role exists and if permission exists

+++ identity-service/identity-domain/identity-application-service/src/main/java/com/tuber/identity/service/domain/helper/role/RemoveRoleFromUserHelper.java
@@ -1,13 +1,17 @@
 package com.tuber.identity.service.domain.helper.role;
 
 import com.tuber.application.handler.ApiResponse;
+import com.tuber.identity.service.domain.constant.IdentityResponseCode;
 import com.tuber.identity.service.domain.dto.user.account.RemoveRoleFromUserCommand;
 import com.tuber.identity.service.domain.dto.user.account.RemoveRoleFromUserResponseData;
 import com.tuber.identity.service.domain.entity.Role;
+import com.tuber.identity.service.domain.ports.output.repository.RoleRepository;
 import com.tuber.identity.service.domain.ports.output.repository.UserAccountRepository;
+import com.tuber.identity.service.domain.exception.IdentityDomainException;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,12 +23,17 @@ import java.util.Set;
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class RemoveRoleFromUserHelper {
     UserAccountRepository userAccountRepository;
+    RoleRepository roleRepository;
 
     @Transactional
     public ApiResponse<RemoveRoleFromUserResponseData> removeRoleFromUser(RemoveRoleFromUserCommand removeRoleFromUserCommand) {
         String username = removeRoleFromUserCommand.getUsername();
+        String roleName = removeRoleFromUserCommand.getRoleName();
+        if (!roleRepository.existsByNameAndUsername(username, roleName)) {
+            throw new IdentityDomainException(IdentityResponseCode.ROLE_NOT_EXISTS, HttpStatus.BAD_REQUEST.value());
+        }
         Set<String> rolesSet = new HashSet<>();
-        rolesSet.add(removeRoleFromUserCommand.getRoleName());
+        rolesSet.add(roleName);
 
         Set<Role> newRolesSet = userAccountRepository.removeRolesFromUser(username, rolesSet);
         return ApiResponse.<RemoveRoleFromUserResponseData>builder()
